pyproject.toml
[project]
name = "sapa-speaker-tracker"
version = "0.1.0"
description = "MCP server for managing SAPA Data Science community speaker database via Notion"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.6.0",
    "notion-client>=2.2.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sapa_speaker_tracker"]
